Update policies for all models

+++ app/controllers/claims_controller.rb
@@ -3,6 +3,7 @@ class ClaimsController < ApplicationController
   def create
      @item = Item.find(params[:item_id])
      @claim = Claim.create(user: current_user, item: @item)
+     authorize @claim
  
      if @claim.save 
        flash[:notice] = "You have claimed \"#{@item.name}\". Make sure to unclaim this item should anything change."
@@ -16,6 +17,7 @@ class ClaimsController < ApplicationController
   def destroy
     @item = Item.find(params[:item_id])
     @claim = Claim.find(params[:id])
+    authorize @claim
 
     if @claim.destroy
        flash[:notice] = "You have unclaimed \"#{@item.name}\"."
